+++ packages/graphic-interface/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mlightcad/graphic-interface
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @mlightcad/geometry-engine@1.0.5
+
 ## 1.0.3
 
 ### Patch Changes