+++ packages/libdxfrw-converter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mlightcad/libdxfrw-converter
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @mlightcad/data-model@1.0.7
+
 ## 1.0.1
 
 ### Patch Changes